examples/getting-started/conversation/send_handle_incoming_sms/pyproject.toml
[tool.poetry]
name = "sinch-getting-started-send-handle-incoming-sms"
version = "0.1.0"
description = "Getting Started: send and handle incoming SMS with Conversation API (DISPATCH, channel identity)"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
flask = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
